fix: move to original position
