+++ app/src/main/java/org/dhis2/usescases/teiDashboard/ui/TeiDashboardMenu.kt
@@ -1,4 +1,4 @@
-package org.dhis2.usescases.teiDashboard
+package org.dhis2.usescases.teiDashboard.ui
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.Assignment
@@ -16,7 +16,23 @@ import androidx.compose.material.icons.outlined.Timeline
 import androidx.compose.material.icons.outlined.Workspaces
 import org.dhis2.R
 import org.dhis2.commons.resources.ResourceManager
-import org.dhis2.usescases.teiDashboard.ui.addSendSmsMenuItem
+import org.dhis2.usescases.teiDashboard.DashboardEnrollmentModel
+import org.dhis2.usescases.teiDashboard.DashboardViewModel
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.ACTIVATE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.COMPLETE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.DEACTIVATE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.DELETE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.ENROLLMENTS
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.FOLLOW_UP
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.GROUP_BY_STAGE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.HELP
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.REMOVE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.SHARE
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.SYNC
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.TRANSFER
+import org.dhis2.usescases.teiDashboard.EnrollmentMenuItem.VIEW_TIMELINE
+import org.dhis2.usescases.teiDashboard.TeiDashboardContracts.Presenter
 import org.hisp.dhis.android.core.enrollment.EnrollmentStatus
 import org.hisp.dhis.mobile.ui.designsystem.component.menu.MenuItemData
 import org.hisp.dhis.mobile.ui.designsystem.component.menu.MenuItemStyle
@@ -27,7 +43,7 @@ import org.hisp.dhis.mobile.ui.designsystem.theme.TextColor
 fun getEnrollmentMenuList(
     enrollmentUid: String?,
     resourceManager: ResourceManager,
-    presenter: TeiDashboardContracts.Presenter,
+    presenter: Presenter,
     dashboardViewModel: DashboardViewModel,
 ): List<MenuItemData<EnrollmentMenuItem>> {
     return if (enrollmentUid == null) {
@@ -39,7 +55,7 @@ fun getEnrollmentMenuList(
 
 private fun buildMenuForNoEnrollment(
     resourceManager: ResourceManager,
-    presenter: TeiDashboardContracts.Presenter,
+    presenter: Presenter,
 ): List<MenuItemData<EnrollmentMenuItem>> {
     return buildList {
         addSyncMenuItem(resourceManager)
@@ -51,7 +67,7 @@ private fun buildMenuForNoEnrollment(
 private fun buildMenuForEnrollment(
     enrollmentUid: String,
     resourceManager: ResourceManager,
-    presenter: TeiDashboardContracts.Presenter,
+    presenter: Presenter,
     dashboardViewModel: DashboardViewModel,
 ): List<MenuItemData<EnrollmentMenuItem>> {
     return buildList {
@@ -79,7 +95,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addSyncMenuItem(
 ) {
     add(
         MenuItemData(
-            id = EnrollmentMenuItem.SYNC,
+            id = SYNC,
             label = resourceManager.getString(R.string.refresh_this_record),
             leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.Sync),
         ),
@@ -93,7 +109,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addIfTeiCanBeTransferr
     if (dashboardViewModel.checkIfTeiCanBeTransferred()) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.TRANSFER,
+                id = TRANSFER,
                 label = resourceManager.getString(R.string.transfer),
                 leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.MoveDown),
             ),
@@ -108,7 +124,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addFollowUpMenuItem(
     if (!dashboardViewModel.showFollowUpBar.value) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.FOLLOW_UP,
+                id = FOLLOW_UP,
                 label = resourceManager.getString(R.string.mark_follow_up),
                 leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.Flag),
             ),
@@ -123,7 +139,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addTimelineOrGroupBySt
     if (dashboardViewModel.groupByStage.value != false) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.VIEW_TIMELINE,
+                id = VIEW_TIMELINE,
                 label = resourceManager.getString(R.string.view_timeline),
                 leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.Timeline),
             ),
@@ -131,7 +147,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addTimelineOrGroupBySt
     } else {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.GROUP_BY_STAGE,
+                id = GROUP_BY_STAGE,
                 label = resourceManager.getString(R.string.group_by_stage),
                 leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.Workspaces),
             ),
@@ -144,7 +160,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addHelpMenuItem(
 ) {
     add(
         MenuItemData(
-            id = EnrollmentMenuItem.HELP,
+            id = HELP,
             label = resourceManager.getString(R.string.showHelp),
             leadingElement = MenuLeadingElement.Icon(icon = Icons.AutoMirrored.Outlined.HelpOutline),
         ),
@@ -156,7 +172,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addMoreEnrollmentsMenu
 ) {
     add(
         MenuItemData(
-            id = EnrollmentMenuItem.ENROLLMENTS,
+            id = ENROLLMENTS,
             label = resourceManager.getString(R.string.more_enrollments),
             leadingElement = MenuLeadingElement.Icon(icon = Icons.AutoMirrored.Outlined.Assignment),
         ),
@@ -168,7 +184,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addShareMenuItem(
 ) {
     add(
         MenuItemData(
-            id = EnrollmentMenuItem.SHARE,
+            id = SHARE,
             label = resourceManager.getString(R.string.share),
             showDivider = true,
             leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.Share),
@@ -179,13 +195,13 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addShareMenuItem(
 private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addStatusMenuItems(
     enrollmentUid: String,
     resourceManager: ResourceManager,
-    presenter: TeiDashboardContracts.Presenter,
+    presenter: Presenter,
 ) {
     val status = presenter.getEnrollmentStatus(enrollmentUid)
     if (status != EnrollmentStatus.COMPLETED) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.COMPLETE,
+                id = COMPLETE,
                 label = resourceManager.getString(R.string.complete),
                 leadingElement = MenuLeadingElement.Icon(
                     icon = Icons.Outlined.CheckCircle,
@@ -199,7 +215,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addStatusMenuItems(
     if (status != EnrollmentStatus.ACTIVE) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.ACTIVATE,
+                id = ACTIVATE,
                 label = resourceManager.getString(R.string.re_open),
                 showDivider = status == EnrollmentStatus.CANCELLED,
                 leadingElement = MenuLeadingElement.Icon(
@@ -214,7 +230,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addStatusMenuItems(
     if (status != EnrollmentStatus.CANCELLED) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.DEACTIVATE,
+                id = DEACTIVATE,
                 label = resourceManager.getString(R.string.deactivate),
                 showDivider = true,
                 leadingElement = MenuLeadingElement.Icon(
@@ -230,7 +246,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addStatusMenuItems(
 private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addRemoveEnrollmentItem(
     enrollmentUid: String,
     resourceManager: ResourceManager,
-    presenter: TeiDashboardContracts.Presenter,
+    presenter: Presenter,
     dashboardViewModel: DashboardViewModel,
 ) {
     if (presenter.checkIfEnrollmentCanBeDeleted(enrollmentUid)) {
@@ -242,7 +258,7 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addRemoveEnrollmentIte
         }
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.REMOVE,
+                id = REMOVE,
                 label = resourceManager.getString(R.string.remove_from),
                 supportingText = programmeName,
                 style = MenuItemStyle.ALERT,
@@ -253,13 +269,13 @@ private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addRemoveEnrollmentIte
 }
 
 private fun MutableList<MenuItemData<EnrollmentMenuItem>>.addDeleteTeiMenuItem(
-    presenter: TeiDashboardContracts.Presenter,
+    presenter: Presenter,
     resourceManager: ResourceManager,
 ) {
     if (presenter.checkIfTEICanBeDeleted()) {
         add(
             MenuItemData(
-                id = EnrollmentMenuItem.DELETE,
+                id = DELETE,
                 label = resourceManager.getString(R.string.dashboard_menu_delete_tei_v2, presenter.teType),
                 style = MenuItemStyle.ALERT,
                 leadingElement = MenuLeadingElement.Icon(icon = Icons.Outlined.DeleteForever),
